# TSDB: Track count of unknown series referenced during WAL/WBL replay