In this PR:
- track total number of series records referenced by WAL/WBL records with unknown series references
- track tombstones that reference unknown series separately from samples
- add counter metrics for unknown series references during WAL/WBL replay